Remove mod.ts from src/internal
